+++ scm-core/src/main/java/sonia/scm/lifecycle/RestartEvent.java
@@ -29,63 +29,37 @@ package sonia.scm.lifecycle;
 import sonia.scm.event.Event;
 
 /**
- * This event can be used to force a restart of the webapp context. The restart
- * event is useful during plugin development, because we don't have to restart
- * the whole server, to see our changes. The restart event could also be used
- * to install or upgrade plugins.
- *
- * But the restart event should be used carefully, because the whole context
- * will be restarted and that process could take some time.
- *
+ * This event indicates a forced restart of scm-manager.
  * @author Sebastian Sdorra
  * @since 2.0.0
  */
 @Event
-public class RestartEvent
-{
+public class RestartEvent {
 
-  /**
-   * Constructs ...
-   *
-   *
-   * @param cause
-   * @param reason
-   */
-  public RestartEvent(Class<?> cause, String reason)
-  {
+  private final Class<?> cause;
+  private final String reason;
+
+  RestartEvent(Class<?> cause, String reason) {
     this.cause = cause;
     this.reason = reason;
   }
 
-  //~--- get methods ----------------------------------------------------------
-
   /**
    * The class which has fired the restart event.
    *
-   *
    * @return class which has fired the restart event
    */
-  public Class<?> getCause()
-  {
+  public Class<?> getCause() {
     return cause;
   }
 
   /**
    * Returns the reason for the restart.
    *
-   *
    * @return reason for restart
    */
-  public String getReason()
-  {
+  public String getReason() {
     return reason;
   }
 
-  //~--- fields ---------------------------------------------------------------
-
-  /** cause of restart */
-  private final Class<?> cause;
-
-  /** reason for restart */
-  private final String reason;
 }