+++ scm-ui/ui-components/src/layout/SecondaryNavigationColumn.tsx
@@ -21,18 +21,39 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
+import React, { ReactNode } from "react";
+import styled from "styled-components";
 
-package sonia.scm.lifecycle.classloading;
+type Props = {
+  children?: ReactNode;
+  collapsed: boolean;
+};
 
-import org.junit.jupiter.api.Test;
+const SecondaryColumn = styled.div<{ collapsed: boolean }>`
+  /* In Bulma there is unfortunately no intermediate step between .is-1 and .is-2, hence the size. 
+  Navigation size should be as constant as possible. */
+  flex: none;
+  width: ${props => (props.collapsed ? "5.5rem" : "20.5rem")};
+  max-width: ${(props: { collapsed: boolean }) => (props.collapsed ? "11.3%" : "25%")};
+  /* Render this column to full size if column construct breaks (page size too small). */
+  @media (max-width: 785px) {
+    width: 100%;
+    max-width: 100%;
+  }
+`;
 
-import static org.assertj.core.api.Assertions.assertThat;
+export default class SecondaryNavigationColumn extends React.Component<Props> {
+  static defaultProps = {
+    collapsed: false
+  };
 
-class ClassLoaderLifeCycleTest {
+  render() {
+    const { children, collapsed } = this.props;
 
-  @Test
-  void shouldCreateDefaultClassLoader() {
-    ClassLoaderLifeCycle classLoaderLifeCycle = ClassLoaderLifeCycle.create();
-    assertThat(classLoaderLifeCycle).isInstanceOf(SimpleClassLoaderLifeCycle.class);
+    return (
+      <SecondaryColumn className="column" collapsed={collapsed}>
+        {children}
+      </SecondaryColumn>
+    );
   }
 }