+++ scm-ui/ui-components/src/layout/index.ts
@@ -31,3 +31,7 @@ export { default as Page } from "./Page";
 export { default as PageActions } from "./PageActions";
 export { default as Subtitle } from "./Subtitle";
 export { default as Title } from "./Title";
+export { default as CustomQueryFlexWrappedColumns } from "./CustomQueryFlexWrappedColumns";
+export { default as PrimaryContentColumn } from "./PrimaryContentColumn";
+export { default as SecondaryNavigationColumn } from "./SecondaryNavigationColumn";
+