+++ scm-ui/ui-webapp/src/admin/plugins/components/ExecutePendingModal.tsx
@@ -27,6 +27,7 @@ import { PendingPlugins } from "@scm-manager/ui-types";
 import { WithTranslation, withTranslation } from "react-i18next";
 import waitForRestart from "./waitForRestart";
 import SuccessNotification from "./SuccessNotification";
+import PendingPluginsQueue from "./PendingPluginsQueue";
 
 type Props = WithTranslation & {
   onClose: () => void;
@@ -85,70 +86,14 @@ class ExecutePendingModal extends React.Component<Props, State> {
       });
   };
 
-  renderInstallQueue = () => {
-    const { pendingPlugins, t } = this.props;
-    return (
-      <>
-        {pendingPlugins._embedded && pendingPlugins._embedded.new.length > 0 && (
-          <>
-            <strong>{t("plugins.modal.installQueue")}</strong>
-            <ul>
-              {pendingPlugins._embedded.new.map(plugin => (
-                <li key={plugin.name}>{plugin.name}</li>
-              ))}
-            </ul>
-          </>
-        )}
-      </>
-    );
-  };
-
-  renderUpdateQueue = () => {
-    const { pendingPlugins, t } = this.props;
-    return (
-      <>
-        {pendingPlugins._embedded && pendingPlugins._embedded.update.length > 0 && (
-          <>
-            <strong>{t("plugins.modal.updateQueue")}</strong>
-            <ul>
-              {pendingPlugins._embedded.update.map(plugin => (
-                <li key={plugin.name}>{plugin.name}</li>
-              ))}
-            </ul>
-          </>
-        )}
-      </>
-    );
-  };
-
-  renderUninstallQueue = () => {
-    const { pendingPlugins, t } = this.props;
-    return (
-      <>
-        {pendingPlugins._embedded && pendingPlugins._embedded.uninstall.length > 0 && (
-          <>
-            <strong>{t("plugins.modal.uninstallQueue")}</strong>
-            <ul>
-              {pendingPlugins._embedded.uninstall.map(plugin => (
-                <li key={plugin.name}>{plugin.name}</li>
-              ))}
-            </ul>
-          </>
-        )}
-      </>
-    );
-  };
-
   renderBody = () => {
-    const { t } = this.props;
+    const { pendingPlugins, t } = this.props;
     return (
       <>
         <div className="media">
           <div className="content">
             <p>{t("plugins.modal.executePending")}</p>
-            {this.renderInstallQueue()}
-            {this.renderUpdateQueue()}
-            {this.renderUninstallQueue()}
+            <PendingPluginsQueue pendingPlugins={pendingPlugins} />
           </div>
         </div>
         <div className="media">{this.renderNotifications()}</div>