+++ scm-webapp/src/main/java/sonia/scm/api/v2/resources/PluginDtoMapper.java
@@ -21,12 +21,13 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-    
+
 package sonia.scm.api.v2.resources;
 
 import de.otto.edison.hal.Links;
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingTarget;
+import sonia.scm.lifecycle.Restarter;
 import sonia.scm.plugin.AvailablePlugin;
 import sonia.scm.plugin.InstalledPlugin;
 import sonia.scm.plugin.Plugin;
@@ -46,6 +47,9 @@ public abstract class PluginDtoMapper {
 
   @Inject
   private ResourceLinks resourceLinks;
+
+  @Inject
+  private Restarter restarter;
 
   public abstract void map(PluginInformation plugin, @MappingTarget PluginDto dto);
 
@@ -78,10 +82,18 @@ public abstract class PluginDtoMapper {
         .self(information.getName()));
 
     if (!plugin.isPending() && PluginPermissions.manage().isPermitted()) {
-      links.single(link("install", resourceLinks.availablePlugin().install(information.getName())));
+      String href = resourceLinks.availablePlugin().install(information.getName());
+      appendLink(links, "install", href);
     }
 
     return new PluginDto(links.build());
+  }
+
+  private void appendLink(Links.Builder links, String name, String href) {
+    links.single(link(name, href));
+    if (restarter.isSupported()) {
+      links.single(link(name + "WithRestart", href + "?restart=true"));
+    }
   }
 
   private PluginDto createDtoForInstalled(InstalledPlugin plugin, List<AvailablePlugin> availablePlugins) {
@@ -96,13 +108,16 @@ public abstract class PluginDtoMapper {
       && !availablePlugin.get().isPending()
       && PluginPermissions.manage().isPermitted()
     ) {
-      links.single(link("update", resourceLinks.availablePlugin().install(information.getName())));
+      String href = resourceLinks.availablePlugin().install(information.getName());
+      appendLink(links, "update", href);
     }
+
     if (plugin.isUninstallable()
       && (!availablePlugin.isPresent() || !availablePlugin.get().isPending())
       && PluginPermissions.manage().isPermitted()
     ) {
-      links.single(link("uninstall", resourceLinks.installedPlugin().uninstall(information.getName())));
+      String href = resourceLinks.installedPlugin().uninstall(information.getName());
+      appendLink(links, "uninstall", href);
     }
 
     PluginDto dto = new PluginDto(links.build());