+++ scm-webapp/src/main/java/sonia/scm/lifecycle/CLibrary.java
@@ -21,18 +21,25 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
+package sonia.scm.lifecycle;
 
-package sonia.scm.lifecycle.classloading;
+/**
+ * Interface for native c library.
+ */
+@SuppressWarnings({
+  "squid:S1214", // usage as constant is common practice for jna
+  "squid:S1191"  // use of sun.* classes is required for jna
+})
+interface CLibrary extends com.sun.jna.Library {
+  CLibrary LIBC =  com.sun.jna.Native.load("c", CLibrary.class);
 
-import org.junit.jupiter.api.Test;
+  int F_GETFD = 1;
+  int F_SETFD = 2;
+  int FD_CLOEXEC = 1;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-class ClassLoaderLifeCycleTest {
-
-  @Test
-  void shouldCreateDefaultClassLoader() {
-    ClassLoaderLifeCycle classLoaderLifeCycle = ClassLoaderLifeCycle.create();
-    assertThat(classLoaderLifeCycle).isInstanceOf(SimpleClassLoaderLifeCycle.class);
-  }
+  int getdtablesize();
+  int fcntl(int fd, int command);
+  int fcntl(int fd, int command, int flags);
+  int execvp(String file, com.sun.jna.StringArray args);
+  String strerror(int errno);
 }